requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.0
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.0
fastapi==0.115.5
ffmpeg-python==0.2.0
filelock==3.12.2
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
greenlet==3.1.1
h11==0.13.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.26.2
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.8
llvmlite==0.43.0
lockfile==0.12.2
loguru==0.7.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.0
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.13.4
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai-whisper==20230124
optuna==4.1.0
pandas==2.2.3
pillow==11.0.0
pooch==1.8.2
primePy==1.3
propcache==0.2.0
protobuf==5.28.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pymongo==3.5.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
stable-ts==1.1.1b0
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1+cu118
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu118
torchmetrics==1.6.0
torchvision==0.20.1+cu118
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
typer==0.13.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==6.0.0
yarl==1.17.1
